Moved the functions to single file
