+++ llvm/lib/Transforms/PeepholeOptimizationCourse/CMakeLists.txt
@@ -2,3 +2,4 @@ add_subdirectory(RemoveAddWithZero)
 add_subdirectory(RemoveAddWithZeroNewPassManager)
 add_subdirectory(ObfuscateI8Add)
 add_subdirectory(ObfuscateI8AddNewPassManager)
+add_subdirectory(ObfuscateI8AddFunctions)
