+++ llvm/lib/Transforms/PeepholeOptimizationCourse/ObfuscateI8AddFunctions/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_llvm_component_library(
+    LLVMObfuscateI8AddFunctions
+    ObfuscateI8AddFunctions.cpp
+)
